backend/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
websockets==13.1
httpx==0.27.2
# ML — anomaly detectors (merged from detection-service)
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
# CPU-only torch keeps the image small; Docker build pulls from the CPU index.
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
# MLOps
wandb>=0.16.0
